AI/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.0
pydantic==2.10.3
requests==2.32.3
Pillow==11.0.0
torch
torchvision
transformers==4.46.2
numpy==1.26.4
grpcio==1.67.0
